requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
distro==1.8.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==5.2.0
imutils==0.5.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
scikit-build==0.16.4
scipy==1.9.3
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
uWSGI==2.0.18
